+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Drop support for Ruby 3.0. The minimum required Ruby version is now 3.1.4.
 * Add support for Ruby 3.3.
+* Allow acronyms within example blocks.
 
 ## 8.3.4