+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 8.4.0
 
 * Drop support for Ruby 3.0. The minimum required Ruby version is now 3.1.4.
 * Add support for Ruby 3.3.